Refactor update-webhook code/unit-tests into operatorinit package
