+++ platform-operator/internal/k8s/certificate/certificate.go
@@ -12,19 +12,16 @@ import (
 	"crypto/x509/pkix"
 	"encoding/pem"
 	"fmt"
-	"go.uber.org/zap"
-	v1 "k8s.io/api/core/v1"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apiextensionsv1client "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"math/big"
 	"os"
 	"time"
 
-	adminv1 "k8s.io/api/admissionregistration/v1"
+	"go.uber.org/zap"
+	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
+	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
 const (
@@ -38,8 +35,8 @@ const (
 	OperatorNamespace = "verrazzano-install"
 	CRDName           = "verrazzanos.install.verrazzano.io"
 
-	certKey = "tls.crt"
-	privKey = "tls.key"
+	CertKey = "tls.crt"
+	PrivKey = "tls.key"
 
 	certYearsValid = 3
 )
@@ -64,12 +61,12 @@ func CreateWebhookCertificates(log *zap.SugaredLogger, kubeClient kubernetes.Int
 		return err
 	}
 
-	if err := writeFile(log, fmt.Sprintf("%s/%s", certDir, certKey), serverPEM); err != nil {
+	if err := writeFile(log, fmt.Sprintf("%s/%s", certDir, CertKey), serverPEM); err != nil {
 		log.Errorf("Error writing cert file: %v", err)
 		return err
 	}
 
-	if err := writeFile(log, fmt.Sprintf("%s/%s", certDir, privKey), serverKeyPEM); err != nil {
+	if err := writeFile(log, fmt.Sprintf("%s/%s", certDir, PrivKey), serverKeyPEM); err != nil {
 		log.Errorf("Error 3 %v", err)
 		return err
 	}
@@ -82,7 +79,7 @@ func createTLSCert(log *zap.SugaredLogger, kubeClient kubernetes.Interface, comm
 	existingSecret, err := secretsClient.Get(context.TODO(), OperatorTLS, metav1.GetOptions{})
 	if err == nil {
 		log.Infof("Secret %s exists, using...", OperatorTLS)
-		return existingSecret.Data[certKey], existingSecret.Data[privKey], nil
+		return existingSecret.Data[CertKey], existingSecret.Data[PrivKey], nil
 	}
 	if !errors.IsNotFound(err) {
 		return []byte{}, []byte{}, err
@@ -148,8 +145,8 @@ func createTLSCertSecretIfNecesary(log *zap.SugaredLogger, secretsClient corev1.
 	webhookCrt.Name = OperatorTLS
 	webhookCrt.Type = v1.SecretTypeTLS
 	webhookCrt.Data = make(map[string][]byte)
-	webhookCrt.Data[certKey] = serverPEMBytes
-	webhookCrt.Data[privKey] = serverKeyPEMBytes
+	webhookCrt.Data[CertKey] = serverPEMBytes
+	webhookCrt.Data[PrivKey] = serverKeyPEMBytes
 
 	_, createError := secretsClient.Create(context.TODO(), &webhookCrt, metav1.CreateOptions{})
 	if createError != nil {
@@ -160,7 +157,7 @@ func createTLSCertSecretIfNecesary(log *zap.SugaredLogger, secretsClient corev1.
 				return []byte{}, []byte{}, err
 			}
 			log.Infof("Secret %s exists, using...", OperatorTLS)
-			return existingSecret.Data[certKey], existingSecret.Data[privKey], nil
+			return existingSecret.Data[CertKey], existingSecret.Data[PrivKey], nil
 		}
 		return []byte{}, []byte{}, createError
 	}
@@ -225,8 +222,8 @@ func createCACertSecretIfNecessary(log *zap.SugaredLogger, secretsClient corev1.
 	webhookCA.Type = v1.SecretTypeTLS
 
 	webhookCA.Data = make(map[string][]byte)
-	webhookCA.Data[certKey] = caPEMBytes
-	webhookCA.Data[privKey] = caKeyPEMBytes
+	webhookCA.Data[CertKey] = caPEMBytes
+	webhookCA.Data[PrivKey] = caKeyPEMBytes
 
 	_, createError := secretsClient.Create(context.TODO(), &webhookCA, metav1.CreateOptions{})
 	if createError != nil {
@@ -264,11 +261,11 @@ func encodeCABytes(caBytes []byte, caPrivKey *rsa.PrivateKey) ([]byte, []byte) {
 
 // decodeExistingSecretData Decode existing secret data into their X509 and RSA objects
 func decodeExistingSecretData(secret *v1.Secret) (*x509.Certificate, *rsa.PrivateKey, error) {
-	cert, err := decodeCertificate(secret.Data[certKey])
+	cert, err := decodeCertificate(secret.Data[CertKey])
 	if err != nil {
 		return nil, nil, err
 	}
-	key, err := decodeKey(secret.Data[privKey])
+	key, err := decodeKey(secret.Data[PrivKey])
 	if err != nil {
 		return nil, nil, err
 	}
@@ -321,95 +318,5 @@ func writeFile(log *zap.SugaredLogger, filepath string, pemData []byte) error {
 		return err
 	}
 
-	return nil
-}
-
-// DeleteValidatingWebhookConfiguration deletes a validating webhook configuration
-func DeleteValidatingWebhookConfiguration(kubeClient kubernetes.Interface, name string) error {
-	_, err := kubeClient.AdmissionregistrationV1().ValidatingWebhookConfigurations().Get(context.TODO(), name, metav1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			return nil
-		}
-		return err
-	}
-	return kubeClient.AdmissionregistrationV1().ValidatingWebhookConfigurations().Delete(context.TODO(), name, metav1.DeleteOptions{})
-}
-
-// UpdateValidatingWebhookConfiguration sets the CABundle
-func UpdateValidatingWebhookConfiguration(kubeClient kubernetes.Interface, name string) error {
-	validatingWebhook, err := kubeClient.AdmissionregistrationV1().ValidatingWebhookConfigurations().Get(context.TODO(), name, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	caSecret, errX := kubeClient.CoreV1().Secrets(OperatorNamespace).Get(context.TODO(), OperatorCA, metav1.GetOptions{})
-	if errX != nil {
-		return errX
-	}
-	crt := caSecret.Data[certKey]
-
-	for i := range validatingWebhook.Webhooks {
-		validatingWebhook.Webhooks[i].ClientConfig.CABundle = crt
-	}
-
-	_, err = kubeClient.AdmissionregistrationV1().ValidatingWebhookConfigurations().Update(context.TODO(), validatingWebhook, metav1.UpdateOptions{})
-	return err
-}
-
-// UpdateConversionWebhookConfiguration sets the conversion webhook for the Verrazzano resource
-func UpdateConversionWebhookConfiguration(apiextClient apiextensionsv1client.ApiextensionsV1Interface, kubeClient kubernetes.Interface) error {
-	crd, err := apiextClient.CustomResourceDefinitions().Get(context.TODO(), CRDName, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	convertPath := "/convert"
-	var webhookPort int32 = 443
-	caSecret, err := kubeClient.CoreV1().Secrets(OperatorNamespace).Get(context.TODO(), OperatorCA, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	crt := caSecret.Data[certKey]
-
-	crd.Spec.Conversion = &apiextensionsv1.CustomResourceConversion{
-		Strategy: apiextensionsv1.WebhookConverter,
-		Webhook: &apiextensionsv1.WebhookConversion{
-			ClientConfig: &apiextensionsv1.WebhookClientConfig{
-				Service: &apiextensionsv1.ServiceReference{
-					Name:      OperatorName,
-					Namespace: OperatorNamespace,
-					Path:      &convertPath,
-					Port:      &webhookPort,
-				},
-				CABundle: crt,
-			},
-			ConversionReviewVersions: []string{"v1beta1"},
-		},
-	}
-	_, err = apiextClient.CustomResourceDefinitions().Update(context.TODO(), crd, metav1.UpdateOptions{})
-	return err
-}
-
-// UpdateMutatingWebhookConfiguration sets the CABundle
-func UpdateMutatingWebhookConfiguration(kubeClient kubernetes.Interface, name string) error {
-	var webhook *adminv1.MutatingWebhookConfiguration
-	webhook, err := kubeClient.AdmissionregistrationV1().MutatingWebhookConfigurations().Get(context.TODO(), name, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	caSecret, err := kubeClient.CoreV1().Secrets(OperatorNamespace).Get(context.TODO(), OperatorCA, metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	crt := caSecret.Data[certKey]
-	if err != nil {
-		return err
-	}
-	for i := range webhook.Webhooks {
-		webhook.Webhooks[i].ClientConfig.CABundle = crt
-	}
-	_, err = kubeClient.AdmissionregistrationV1().MutatingWebhookConfigurations().Update(context.TODO(), webhook, metav1.UpdateOptions{})
-	if err != nil {
-		return err
-	}
 	return nil
 }
