+++ platform-operator/internal/operatorinit/run_webhook.go
@@ -150,13 +150,13 @@ func setupWebhooksWithManager(log *zap.SugaredLogger, mgr manager.Manager, kubeC
 // updateWebhookConfigurations Creates or updates the webhook configurations as needed
 func updateWebhookConfigurations(kubeClient *kubernetes.Clientset, log *zap.SugaredLogger, conf *rest.Config) error {
 	log.Debug("Delete old VPO webhook configuration")
-	if err := certificate.DeleteValidatingWebhookConfiguration(kubeClient, certificate.OldOperatorName); err != nil {
+	if err := deleteValidatingWebhookConfiguration(kubeClient, certificate.OldOperatorName); err != nil {
 		return fmt.Errorf("Failed to delete old webhook configuration: %v", err)
 	}
 
 	log.Debug("Updating VPO webhook configuration")
 
-	if err := certificate.UpdateValidatingWebhookConfiguration(kubeClient, certificate.OperatorName); err != nil {
+	if err := updateValidatingWebhookConfiguration(kubeClient, certificate.OperatorName); err != nil {
 		return fmt.Errorf("Failed to update validation webhook configuration: %v", err)
 	}
 
@@ -166,16 +166,16 @@ func updateWebhookConfigurations(kubeClient *kubernetes.Clientset, log *zap.Suga
 		return fmt.Errorf("Failed to get apix clientset: %v", err)
 	}
 
-	if err := certificate.UpdateConversionWebhookConfiguration(apixClient, kubeClient); err != nil {
+	if err := updateConversionWebhookConfiguration(apixClient, kubeClient); err != nil {
 		return fmt.Errorf("Failed to update conversion webhook: %v", err)
 	}
 
-	if err := certificate.UpdateMutatingWebhookConfiguration(kubeClient, constants.MysqlBackupMutatingWebhookName); err != nil {
+	if err := updateMutatingWebhookConfiguration(kubeClient, constants.MysqlBackupMutatingWebhookName); err != nil {
 		return fmt.Errorf("Failed to update pod mutating webhook configuration: %v", err)
 	}
 
 	log.Debug("Updating MySQL install values webhook configuration")
-	if err := certificate.UpdateValidatingWebhookConfiguration(kubeClient, webhooks.MysqlInstallValuesWebhook); err != nil {
+	if err := updateValidatingWebhookConfiguration(kubeClient, webhooks.MysqlInstallValuesWebhook); err != nil {
 		return fmt.Errorf("Failed to update validation webhook configuration: %v", err)
 	}
 	return nil
